Check JamPolygon object

Check whether a JamPolygon object is valid.

check_JamPolygon(object)

Arguments

object

JamPolygon object

Details

It requires:

slotName(object) to contain "polygons"

object@polygons to inherit "data.frame"

colnames(object@polygons) to contain c("name", "x", "y")

General guidance for JamPolygon objects:

Empty polygons can be represented as one row in the data.frame which contains NULL or numeric(0) x,y coordinates.

Empty polygons can be the result of intersect_JamPolygon() that finds no suitable intersect.

Note that downstream processing, and plotting, must account for empty polygons and handle or ignore them accordingly.

See also