Draw boundary around grouops of gridtext labels

Draw boundary around grouops of gridtext labels

draw_gridtext_groups(
  g_labels,
  gdf,
  segment_df = NULL,
  do_draw = TRUE,
  verbose = FALSE,
  ...
)

Arguments

g_labels
richtext_grob class as returned by gridtext::richtext_grob().

gdf
data.frame with one row per grob in g_labels, with colnames "x", "y", "overlap_set", "r", "r_unit", "border_col", "box_fill". The rows are grouped using c("x", "y", "overlap_set"). The "r" and "r_unit" values are used to define the radius and radius unit for the rounded corner. The row in gdf with the highest "r" value is the reference row, for the "r" radius, and the "border_col" and "box_fill" values.

Details

This function is a helper function used to take individual labels defined by gridtext::richtext_grob(), and creating grid::roundrectGrob() objects that cover each group of labels with one rounded rectangular boundary.

See also