Convert euler output to polygons

Convert euler output to polygons

eulerr_to_polygon_list(x)

Arguments

x

output from eulerr::euler()

Value

list polygon object with one polygon for each Euler circle or ellipse.

list with polygons for each unique set defined by names(x), where each list contains numeric vectors named "x" and "y".

Details

This function takes the output from eulerr::euler() and converts it to polygons in list format.

See also

Examples

counts <- c(A=1, B=2, `A&B`=3)
x <- eulerr::euler(counts)
polygon_list <- eulerr_to_polygon_list(x)

polygon_areas(polygon_list)
#> $A
#> [1] 3.999335
#> 
#> $B
#> [1] 4.999169
#>