Define venndir curate data.frame

Define venndir curate data.frame

Usage

get_venndir_curate_df(unicode = TRUE, use_color = TRUE, curate_df = NULL, ...)

Arguments

unicode

logical default TRUE, whether to use unicode output values. Optionally unicode=2 will return an alternative set of Unicode values.

use_color

logical default TRUE, whether to apply color, a convenient method to use dark grey.

curate_df

data.frame with optional user-defined values, which are used in place of corresponding default values in the 'from' column.

...

additional arguments are ignored.

Value

data.frame with colnames:

'from' - character sign to match

'color' - character color associated with matched values

'sign' - character replacement of the matched value in 'from'

'hide_singlet' - logical indicating whether to hide the signed count value when the overlap is a singlet, for example "agreement" has no meaning when there is only one set involved.

Details

This function defines the data.frame used for the argument curate_df in venndir(). The purpose is to make it more convenient to review and potentially modify the default values.

See also

Examples

# default
get_venndir_curate_df()
#>         from sign       color hide_singlet
#> 1         -1    ↓ dodgerblue3        FALSE
#> 2          1    ↑   firebrick        FALSE
#> 3          0    -                    FALSE
#> 4 concordant    = dodgerblue3         TRUE
#> 5  agreement    = dodgerblue3         TRUE
#> 6      mixed    X      grey45        FALSE

# alternate Unicode
get_venndir_curate_df(unicode=2)
#>         from sign       color hide_singlet
#> 1         -1    ↓ dodgerblue3        FALSE
#> 2          1    ↑   firebrick        FALSE
#> 3          0    -                    FALSE
#> 4 concordant    ✓ dodgerblue3         TRUE
#> 5  agreement    ✓ dodgerblue3         TRUE
#> 6      mixed    ✕      grey45        FALSE

# non-unicode
get_venndir_curate_df(unicode=FALSE)
#>         from sign       color hide_singlet
#> 1         -1    v dodgerblue3        FALSE
#> 2          1    ^   firebrick        FALSE
#> 3          0    -                    FALSE
#> 4 concordant    = dodgerblue3         TRUE
#> 5  agreement    = dodgerblue3         TRUE
#> 6      mixed    X      grey45        FALSE

# custom curate_df
my_curate_df <- data.frame(from="1", sign="^", color="red")
get_venndir_curate_df(unicode=TRUE, curate_df=my_curate_df)
#>         from sign       color hide_singlet
#> 1          1    ^         red        FALSE
#> 2         -1    ↓ dodgerblue3        FALSE
#> 4          0    -                    FALSE
#> 5 concordant    = dodgerblue3         TRUE
#> 6  agreement    = dodgerblue3         TRUE
#> 7      mixed    X      grey45        FALSE