Position labels outside JamPolygon

Position labels outside JamPolygon

Usage

label_outside_JamPolygon(
  jp,
  which_jp = NULL,
  center = NULL,
  buffer = 0.1,
  center_method = c("bbox", "label"),
  vector_method = c("label", "farthest"),
  segment_method = c("nearest", "vector"),
  distance = 0.1,
  min_degrees = 15,
  relative = TRUE,
  seed = 123,
  debug = FALSE,
  verbose = FALSE,
  ...
)

Arguments

jp

JamPolygon

which_jp

integer or NULL; when which_jp contains one or more integer values, they refer to rows in jp, and each will be analyzed in sequence. When which_jp=NULL then all the polygons in jp will be analyzed in sequence.

center

numeric vector or matrix with two values indicating the center position. When center=NULL then the center is determined using a method defined by center_method.

center_method

character string indicating the method to determine the center:

"label" uses the mean x,y coordinate of all the polygon label positions;

"bbox" uses the mean x,y coordinate of the bounding box that encompasses the polygons.

The effect is to extend outer labels radially around this center point. Using the mean label position with center_method="label" is helpful because it ensures labels are extended in all directions even when most labels are in one upper or lower section of the sp polygons.

distance

numeric value indicating the absolute coordinate distance away from the perimiter of jp to place labels. This value is the buffer for buffer_JamPolygon().

See also