Optimized list to signed incidence matrix

Optimized list to signed incidence matrix

list2im_value(setlist, empty = NULL, do_sparse = TRUE, force_sign = FALSE, ...)

Arguments

setlist
list of vectors

empty
default single value used for empty/missing entries, the default empty=0 uses zero for entries not present. Another alternative is NA.

do_sparse
logical indicating whether to coerce the output to sparse matrix class "ngCMatrix" from the Matrix package.

...
additional arguments are ignored.

coerce_sign
logical indicating whether to coerce numeric vector values to the sign. When coerce_sign=FALSE the vector values are stored directly. When coerce_sign=TRUE the signs of the vector values are stored.

Value

Matrix object that contains signed direction encoded as c(-1, 0, 1) values.

Details

This function converts a list of named vectors into an incidence matrix with value for each entry (row) present in each input list (column). The rows are defined by the vector names, and values are defined by the vector values.

Note that this function will store zero 0 when the input vector value is zero. When this is not the desired behavior, the argument empty can be used to distinguish missing data from data that is zero, for example by setting empty=NA. In this way a value of zero 0 indicates "present but zero", and a value NA indicates "not present at all". This distinction is helpful when comparing entities which are not tested in each scenario. For example if "geneA" is present and the value is 1 in one list; "geneA" is not tested in the second list; therefore the absence of "geneA" of a non-zero value in the second list is not counted as "non-overlapping" because it was not possible for it to have a non-zero value.

See also

Examples

setlist <- make_venn_test(100, 3, do_signed=TRUE)
imv <- list2im_value(setlist);
print(head(imv));

#> 6 x 3 sparse Matrix of class "dgCMatrix"
#>          set_A set_B set_C
#> item_067    -1     .     .
#> item_042     1     .     .
#> item_050     1     .    -1
#> item_043    -1     .     .
#> item_014    -1     .     .
#> item_025     1    -1     .

# convert back to list
im_value2list(imv);

#> $set_A
#> item_067 item_042 item_050 item_043 item_014 item_025 item_090 item_091 
#>       -1        1        1       -1       -1        1       -1       -1 
#> item_069 item_093 item_057 item_009 item_072 item_026 item_007 item_099 
#>       -1        1        1       -1        1        1       -1       -1 
#> item_089 item_083 item_036 item_078 item_097 item_076 item_015 item_032 
#>       -1        1        1        1        1        1       -1        1 
#> item_086 item_084 item_041 item_023 item_027 item_060 item_053 item_079 
#>        1        1       -1        1       -1        1        1       -1 
#> 
#> $set_B
#> item_025 item_069 item_093 item_072 item_089 item_097 item_076 item_027 
#>       -1       -1        1        1       -1       -1        1       -1 
#> item_053 item_096 item_038 item_034 item_063 item_013 item_082 item_021 
#>        1        1       -1       -1        1        1        1       -1 
#> 
#> $set_C
#> item_050 item_090 item_091 item_072 item_086 item_041 item_060 item_079 
#>       -1       -1        1        1        1       -1        1       -1 
#> item_047 item_095 item_016 item_006 item_039 item_031 item_081 
#>        1        1        1       -1        1       -1        1 
#> 

# make a simple character vector list
setlistv <- lapply(setlist, function(i){
   j <- letters[i+3];
   names(j) <- names(i);
   j;
})
imv <- list2im_value(setlistv);
print(head(imv));

#>          set_A set_B set_C
#> item_067 "b"   ""    ""   
#> item_042 "d"   ""    ""   
#> item_050 "d"   ""    "b"  
#> item_043 "b"   ""    ""   
#> item_014 "b"   ""    ""   
#> item_025 "d"   "b"   ""   

im_value2list(imv);

#> $set_A
#> item_067 item_042 item_050 item_043 item_014 item_025 item_090 item_091 
#>      "b"      "d"      "d"      "b"      "b"      "d"      "b"      "b" 
#> item_069 item_093 item_057 item_009 item_072 item_026 item_007 item_099 
#>      "b"      "d"      "d"      "b"      "d"      "d"      "b"      "b" 
#> item_089 item_083 item_036 item_078 item_097 item_076 item_015 item_032 
#>      "b"      "d"      "d"      "d"      "d"      "d"      "b"      "d" 
#> item_086 item_084 item_041 item_023 item_027 item_060 item_053 item_079 
#>      "d"      "d"      "b"      "d"      "b"      "d"      "d"      "b" 
#> 
#> $set_B
#> item_025 item_069 item_093 item_072 item_089 item_097 item_076 item_027 
#>      "b"      "b"      "d"      "d"      "b"      "b"      "d"      "b" 
#> item_053 item_096 item_038 item_034 item_063 item_013 item_082 item_021 
#>      "d"      "d"      "b"      "b"      "d"      "d"      "d"      "b" 
#> 
#> $set_C
#> item_050 item_090 item_091 item_072 item_086 item_041 item_060 item_079 
#>      "b"      "b"      "d"      "d"      "d"      "b"      "d"      "b" 
#> item_047 item_095 item_016 item_006 item_039 item_031 item_081 
#>      "d"      "d"      "d"      "b"      "d"      "b"      "d" 
#>