Text Venn diagram

Text Venn diagram

textvenn(
  setlist,
  sets = seq_along(setlist),
  overlap_type = c("concordance", "overlap", "each", "agreement"),
  set_colors = NULL,
  spacing = 5,
  padding = 1,
  inverse_title = TRUE,
  inverse_counts = FALSE,
  color_by_counts = TRUE,
  return_items = FALSE,
  unicode = TRUE,
  big.mark = ",",
  verbose = FALSE,
  ...
)

Arguments

setlist
list of item vectors; list of vectors named by item; incidence matrix with values c(0, 1) or c(FALSE, TRUE), or c(-1, 0, 1).

sets
integer vector as index to setlist, used to pull out a subset of the list elements. This subset is useful because the set colors are defined for the full setlist, which allows the subset of colors to be consistent for each set.

set_colors
NULL or character vector that contains R-compatible colors. When set_colors is NULL, categorical colors are defined using colorjam::rainbowJam(). When set_colors is defined, the values are recycled to the total number of sets represented by setlist.

spacing, padding
integer values indicating the character spacing and padding around labels printed to the console output.

inverse_title, inverse_counts
logical indicating whether to inverse the color, when inverse_title=TRUE then each Venn set is printed on colored background, when inverse_title=FALSE, each set is printed with colored text with no background color.

return_items
logical passed to signed_overlaps() to indicate whether to include items with each Venn set. This data is returned using invisible(), and is relevant only when storing the return value.

unicode
logical passed to curate_venn_labels() indicating whether the directional label can include special Unicode characters.

big.mark
character passed to format() for numeric labels.

verbose
logical indicating whether to print verbose output.

Value

data.frame returned using invisible(), from the output of signed_overlaps().

Details

This function is a very simple method to print a Venn diagram using text, intended to be displayed using mono-spaced font on an R console.

See also

Examples

setlist <- make_venn_test(n_items=100, do_signed=TRUE)

# two-way Venn by default shows concordance
textvenn(setlist, sets=c(1,2))

#> Error in venn_m[cbind(set_rownums, set_colnums)] <- names(fCounts): replacement has length zero

# without signed directionality use overlap_type="overlap"
textvenn(setlist, sets=c(1,2), overlap_type="overlap")

#> Error in venn_m[cbind(set_rownums, set_colnums)] <- names(fCounts): replacement has length zero

# three-way Venn showing each signed directionality
textvenn(setlist, sets=c(1,2,3), overlap_type="each")

#> Error in venn_m[cbind(set_rownums, set_colnums)] <- names(fCounts): replacement has length zero

# larger number of items
setlist <- make_venn_test(n_items=1000000, sizes=200000, do_signed=TRUE)
# text Venn with directionality
textvenn(setlist, sets=c(1,2,3), "agreement")

#> Error in venn_m[cbind(set_rownums, set_colnums)] <- names(fCounts): replacement has length zero

# basic text Venn with directionality
textvenn(setlist, sets=c(1,2,3), "each")

#> Error in venn_m[cbind(set_rownums, set_colnums)] <- names(fCounts): replacement has length zero