Directional Venn diagram

Directional Venn diagram

venndir(
  setlist,
  sets = seq_along(setlist),
  set_colors = NULL,
  overlap_type = c("concordance", "each", "overlap", "agreement"),
  proportional = FALSE,
  return_items = FALSE,
  display_zero = TRUE,
  font_cex = c(1, 0.8),
  poly_alpha = 0.5,
  label_style = c("basic", "fill", "shaded", "shaded_box", "lite", "lite_box"),
  x_nudge = NULL,
  y_nudge = NULL,
  circle_nudge = NULL,
  unicode = TRUE,
  big.mark = ",",
  do_plot = TRUE,
  ...
)

Arguments

setlist
list of named vectors, whose names represent set items, and whose values represent direction using values c(-1, 0, 1).

sets
integer index of sets in setlist to display in the Venn diagram. This subset is useful when creating a Venn diagram for a subset of a list, because it defines consistent colors across all sets, and uses the appropriate subset of colors in the Venn diagram.

set_colors
character vector of R colors, or NULL to use default colors defined by colorjam::rainbowJam().

overlap_type
character value indicating the type of overlap logic: "each" records each combination of signs; "overlap" disregards the sign and returns any match item overlap; "concordance" represents counts for full agreement, or "mixed" for any inconsistent overlapping direction; "agreement" represents full agreement in direction as "agreement", and "mixed" for any inconsistent direction.

proportional
logical indicating whether the Venn circles are proportionally sized, also known as a Euler diagram. Note that proportionally sized circles are not guaranteed to represent every possible overlap.

return_items
logical indicating whether to return the items within each overlap set.

display_zero
logical indicating whether empty overlaps are labeled with zero 0 when display_zero=TRUE, or are blank when display_zero=FALSE.

font_cex
numeric vector length 2, indicating relative font size for the main label, and directional label, respectively. The default c(1, 0.8) defines the main label with 1x size, and the directional labels with 80% that size.

poly_alpha
numeric value between 0 and 1, indicating the alpha transparency of the polygon backgroun color, where poly_alpha=1 is completely 100% opaque (no transparency), and poly_alpha=0.4 is 40% opaque, therefore 60% transparency.

label_style
character string indicating the style of label to display: "basic" displays text with no background shading or border, "fill" displays text on opaque colored background, "shaded" displays text on partially transparent colored background, "lite" displays text on partially transparent lite background, "lite_box" displays text on lite background with border.

unicode
logical passed to curate_venn_labels() indicating whether the directional label can include special Unicode characters.

big.mark
character passed to format() for numeric labels.

...
additional arguments are passed to render_venndir().

Details

This function displays a Venn diagram, or when proportional=TRUE it displays a Euler diagram, representing counts in each overlap set.

Input data is supplied either as a list or incidence matrix, or signed incidence matrix (whose values indicate direction), and is passed to signed_overlaps() to summarize counts by Venn overlaps.

By default, when input data contains signed direction, the counts include summary counts for the different forms of agreement in direction. See signed_overlaps() for description of each overlap_type, for different methods of summarizing the overlapping directions.

See also

Examples

setlist <- make_venn_test(100, 3);
venndir(setlist)

#> Error in strsplit(names(counts), split = sep): non-character argument

setlist <- make_venn_test(100, 3, do_signed=TRUE);
venndir(setlist)

#> Error in strsplit(names(counts), split = sep): non-character argument

venndir(setlist)

#> Error in strsplit(names(counts), split = sep): non-character argument

venndir(setlist, label_style="basic")

#> Error in strsplit(names(counts), split = sep): non-character argument

venndir(setlist, label_style="fill")

#> Error in strsplit(names(counts), split = sep): non-character argument

venndir(setlist, label_style="shaded")

#> Error in strsplit(names(counts), split = sep): non-character argument

venndir(setlist, label_style="shaded_box")

#> Error in strsplit(names(counts), split = sep): non-character argument

venndir(setlist, label_style="lite")

#> Error in strsplit(names(counts), split = sep): non-character argument

venndir(setlist, label_style="lite_box")

#> Error in strsplit(names(counts), split = sep): non-character argument

# proportional Euler diagram
venndir(setlist,
   proportional=TRUE);

#> Error in strsplit(names(counts), split = sep): non-character argument

# nudge circles and hide the zero
venndir(setlist,
   proportional=TRUE,
   display_zero=FALSE,
   circle_nudge=list(set_C=c(-0.5, 0.1))
)

#> Error in strsplit(names(counts), split = sep): non-character argument

# nudge circles so one overlap is no longer shown
venndir(setlist,
   proportional=TRUE,
   display_zero=FALSE,
   circle_nudge=list(set_C=c(-1.4, 0.1))
)

#> Error in strsplit(names(counts), split = sep): non-character argument

setlist2k <- make_venn_test(2000, 3, 80, do_signed=TRUE);
venndir(setlist2k)

#> Error in strsplit(names(counts), split = sep): non-character argument

venndir(setlist2k, proportional=TRUE)

#> Error in strsplit(names(counts), split = sep): non-character argument