Curate Venn labels

Curate Venn labels

curate_venn_labels(
  x,
  type = c("sign", "color"),
  curate_df = NULL,
  unicode = TRUE,
  blend_preset = "ryb",
  ...
)

Arguments

x
vector of overlap labels.

type
character string where type="sign" will curate x into directional sign, and type="color" will curate x into corresponding directional color.

curate_df
data.frame or NULL with optional curation rules. The colnames are expected to contain c("from", "sign", "color"). In this case, each row of curate_df is applied in order, and gsub() is called on each "from" and replaced using the value in column given by type.

unicode
logical indicating whether to use Unicode characters when type="sign". Note this argument only affects the default values, it is not applied when using a custom curate_df.

blend_preset
character string passed as preset to colorjam::blend_colors() to define the color wheel used during color blending operations.

...
additional arguments are ignored.

Value

vector of labels or colors, based upon argument type.

Details

This venndir utility function is used to convert a basic directional label such as "0 1 0 -1" suitable for display. It can output either Unicode or non-Unicode text label, or a corresponding color.

The input currently recognizes directional labels such as "0 -1 1 0" or the character labels "agreement", "concordant", "mixed". Note that zeros "0" are typically removed before calling this function.

See also

Examples

venn_labels <- c("0 1 0 -1", "1 -1", "1 1 1", "mixed", "agreement", "1 1 0 0");
(curate_venn_labels(venn_labels, "sign"))

#> [1] "-↑-↓" "↑↓"   "↑↑↑"  "X"    "="    "↑↑--"

(curate_venn_labels(venn_labels, "sign", unicode=FALSE))

#> [1] "-^-v" "^v"   "^^^"  "X"    ">>>"  "^^--"

(curate_venn_labels(venn_labels, "color"))

#> [1] "#954E88FF" "#954E88FF" "#B22222FF" "grey45"    "#1874CDFF" "#B22222FF"

jamba::printDebug(as.list(curate_venn_labels(venn_labels, "sign")),
   collapse=", ",
   fgText=as.list(curate_venn_labels(venn_labels, "color")))

#> Warning: no non-missing arguments to min; returning Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to min; returning Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> ##  (17:18:41) 07Oct2020:  -↑-↓, ↑↓, ↑↑↑, X, =, ↑↑--