venndir label style

venndir label style

venndir_label_style(
  venndir_output,
  label_preset = c("none", "main inside", "main outside", "outside", "all outside",
    "concept", "meme", "items", "main items", "main count items"),
  label_style = c("basic", "box", "fill", "fill_box", "shaded", "shaded_box", "lite",
    "lite_box"),
  lite = "#FFEEAABB",
  bg = "white",
  set = c("inside", "outside", "none"),
  overlap = c("none", "inside", "outside"),
  count = c("inside", "outside", "none"),
  signed = c("inside", "outside", "none"),
  items = c("none", "inside"),
  inside_percent_threshold = 5,
  label_types = c("main", "signed"),
  show_zero = FALSE,
  sep = "&",
  useGrey = 15,
  verbose = FALSE,
  ...
)

Arguments

label_style
character string indicating the style of label to display. The values "basic","none","shaded","lite","fill" style the label background fill, while presence of "box" in the string will draw a border around the label: "basic" or "none" uses no background fill, "lite" uses lite background fill, "fill" uses opaque fill with the overlap set color, "shaded" uses slightly transparent fill with overlap set color, "box" displays border around the label.

Details

This function applies a label style to venndir() output, overwriting the existing label style as necessary.

This function adjusts the label text color for contrast with make_color_contrast(), which is useful when positioning the label on top of dark or bright colors. If the label is positioned outside the sp polygons, the text is assumed to be on a white background, using argument bg.

See also